move non-shared components to directory in `src/screens`
