+++ src/components/discover/DiscoverSearchInput.js
@@ -25,7 +25,7 @@ import styled from '@/styled-thing';
 import { margin, padding } from '@/styles';
 import { deviceUtils, ethereumUtils } from '@/utils';
 import networkInfo from '@/helpers/networkInfo';
-import DiscoverSheetContext from '@/components/discover-screen/DiscoverScreenContext';
+import DiscoverSheetContext from '@/screens/discover/context/DiscoverScreenContext';
 
 export const ExchangeSearchHeight = 40;
 const ExchangeSearchWidth = deviceUtils.dimensions.width - 30;
