+++ src/components/discover/EdgeFade.js
