+++ src/components/expanded-state/asset/SocialLinks.js
@@ -2,7 +2,7 @@ import lang from 'i18n-js';
 import { startCase } from 'lodash';
 import React from 'react';
 import Link from '../../Link';
-import EdgeFade from '../../discover-screen/EdgeFade';
+import EdgeFade from '../../discover/EdgeFade';
 import styled from '@/styled-thing';
 import { ethereumUtils } from '@/utils';
