+++ src/screens/discover/components/DiscoverScreenContent.js
@@ -1,6 +1,6 @@
 import React, { useRef, useState } from 'react';
 import { View } from 'react-native';
-import { FlexItem } from '../layout';
+import { FlexItem } from '../../../components/layout';
 import DiscoverHome from './DiscoverHome';
 import DiscoverSearch from './DiscoverSearch';
 import DiscoverSearchContainer from './DiscoverSearchContainer';
