+++ src/screens/discover/components/DiscoverSearch.js
@@ -12,13 +12,13 @@ import { InteractionManager, View } from 'react-native';
 import { IS_TESTING } from 'react-native-dotenv';
 import { useDispatch } from 'react-redux';
 import { useDebounce } from 'use-debounce/lib';
-import CurrencySelectionTypes from '../../helpers/currencySelectionTypes';
-import { emitAssetRequest } from '../../redux/explorer';
-import deviceUtils from '../../utils/deviceUtils';
-import { CurrencySelectionList } from '../exchange';
-import { initialChartExpandedStateSheetHeight } from '../expanded-state/asset/ChartExpandedState';
-import { Row } from '../layout';
-import DiscoverSheetContext from './DiscoverScreenContext';
+import CurrencySelectionTypes from '@rainbow-me/helpers/currencySelectionTypes';
+import { emitAssetRequest } from '@rainbow-me/redux/explorer';
+import deviceUtils from '@rainbow-me/utils/deviceUtils';
+import { CurrencySelectionList } from '../../../components/exchange';
+import { initialChartExpandedStateSheetHeight } from '../../../components/expanded-state/asset/ChartExpandedState';
+import { Row } from '../../../components/layout';
+import DiscoverSheetContext from '../context/DiscoverScreenContext';
 import { analytics } from '@/analytics';
 import { PROFILES, useExperimentalFlag } from '@/config';
 import { fetchSuggestions } from '@/handlers/ens';
@@ -28,7 +28,7 @@ import {
   useSwapCurrencyList,
 } from '@/hooks';
 import { useNavigation } from '@/navigation';
-import Routes from '@/navigation/routesNames';
+import Routes from '@rainbow-me/routes';
 import styled from '@/styled-thing';
 import { useTheme } from '@/theme';
 import { ethereumUtils } from '@/utils';
