+++ src/screens/discover/components/DiscoverSearchContainer.js
@@ -10,11 +10,11 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { ButtonPressAnimation } from '../animations';
-import { Column, Row } from '../layout';
-import { Text } from '../text';
-import DiscoverSearchInput from './DiscoverSearchInput';
-import DiscoverSheetContext from './DiscoverScreenContext';
+import { ButtonPressAnimation } from '../../../components/animations';
+import { Column, Row } from '../../../components/layout';
+import { Text } from '../../../components/text';
+import DiscoverSearchInput from '../../../components/discover/DiscoverSearchInput';
+import DiscoverSheetContext from '../context/DiscoverScreenContext';
 import { deviceUtils } from '@/utils';
 import { analytics } from '@/analytics';
 import { useDelayedValueWithLayoutAnimation } from '@/hooks';
